+++ core/dataflow/builders.py
@@ -77,8 +77,7 @@ class DagBuilder(abc.ABC):
         """
 
     @abc.abstractmethod
-    # TODO(*): Consider getting rid of the `dag` argument here.
-    def get_dag(self, config: cconfig.Config, dag: Optional[DAG] = None) -> DAG:
+    def get_dag(self, config: cconfig.Config, mode: str = "strict") -> DAG:
         """
         Build DAG given `config`.
