+++ helpers/test/Test_dassert_issubclass1.test_fail1/output/test.txt
@@ -1,0 +1,1 @@
+issubclass() arg 2 must be a class or tuple of classes
