+++ helpers/test/Test_dassert_issubclass1.test_man_fail2/output/test.txt
@@ -1,0 +1,5 @@
+
+################################################################################
+* Failed assertion *
+instance '<helpers.test.test_dbg._Man object at 0x>' of class '_Man' is not a subclass of '<class 'int'>'
+################################################################################
