+++ helpers/unit_test.py
@@ -279,7 +279,9 @@ def to_string(var: str) -> str:
 
 
 def get_random_df(
-    num_cols: int, seed: Optional[int] = None, date_range_kwargs: Optional[Dict[str, Any]] = None
+    num_cols: int,
+    seed: Optional[int] = None,
+    date_range_kwargs: Optional[Dict[str, Any]] = None,
 ) -> "pd.DataFrame":
     """
     Compute df with random data with `num_cols` columns and index obtained by
